Unify CMake code related to enabling language bindings
- add macro  macro_optional_add_bindings(<component_found> <component_name> 
<subdir1> [<subdir2> ...])
# Conditionally enable bindings.
# Pass -DDISABLE_<component_name>=ON to disable certain binding even if its 
dependencies are found.
# CMake project needs to be set before using this macro.
# Macro sets following variables:
#     ${CMAKE_PROJECT_NAME}_ENABLED - list of enabled bindings
#     ${CMAKE_PROJECT_NAME}_DISABLED - list of disabled bindings
# Usage:
#     macro_optional_add_bindings(<component_found> <component_name> <subdir1> 
[<subdir2> ...])
project() needs to be set (preferably in the same scope as macro invocations)

- add macro macro_display_bindings_log()
# Show bindings summary
# Usage:
#     macro_display_bindings_log()
Simple bindings summary, feel free to extend it in a future if you like.

- use them in Smoke, Ruby, CSharp
Unified component names in all bindings as well, so that for instance -DDISABLE_QtWebkit=ON will 
disable WebKit bindings for all languages.

CCMAIL: kde-bindings@kde.org
CCMAIL: kde-buildsystem@kde.org

svn path=/trunk/KDE/kdebindings/ruby/; revision=1125668